dependent_disentanglement/requirements.txt
argparse==1.4.0
numpy==1.21.0
pathlib==1.0.1
h5py==3.6.0
pyYAML==6.0
sklearn==0.0
einops==0.5.0
seaborn==0.12.2
timm==0.8.0.dev0
asteroid==0.6.0
wandb==0.19.0
pymp==0.0.6
transformers==4.42.3
fairscale==0.4.12
opencv-python==4.10.0.84
scikit_image==0.21.0
pyitlib==0.3.0
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.12.1+cu113
torchvision==0.13.1+cu113